clean up custom API
