+++ ihp.cabal
@@ -226,6 +226,7 @@ library
         , IHP.Job.Dashboard.Types
         , IHP.Job.Dashboard.Auth
         , IHP.Job.Dashboard.View
+        , IHP.Job.Dashboard.Utils
         , IHP.PageTitle.ControllerFunctions
         , IHP.PageTitle.Types
         , IHP.PageTitle.ViewFunctions
